feat: implementation CRUD functionality for products

+++ entity/product_entity.go
@@ -6,7 +6,7 @@ type Product struct {
 	ID        uint      `gorm:"primaryKey" json:"id"`
 	Name      string    `gorm:"type:varchar(100);not null" json:"name"`
 	Stock     int       `gorm:"not null" json:"stock"`
-	Price     float64   `gorm:"not null" json:"price"`
+	Price     int64     `gorm:"not null" json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
